some fixes to npm package and readme

+++ npm/installer.js
@@ -7,7 +7,8 @@ const { stdout } = require("node:process");
 const { execFileSync } = require("node:child_process");
 
 const tar = require("tar-fs");
-const unzipper = require("unzipper");
+const { createHash } = require("node:crypto");
+const AdmZip = require("adm-zip");
 
 const ARCH_MAPPING = {
   ia32: "386",
@@ -16,7 +17,6 @@ const ARCH_MAPPING = {
 };
 
 const PLATFORM_MAPPING = {
-  darwin: "darwin",
   linux: "linux",
   win32: "windows",
 };
@@ -44,21 +44,23 @@ async function getCurrentRelease() {
   return res.json();
 }
 
-async function extractFile(path) {
+async function extractFileZip(path, destination) {
+  const zip = new AdmZip(path)
+  return new Promise((resolve, reject) => zip.extractAllToAsync(destination, true, false, (err) => (err ? reject(err) : resolve())));
+}
+
+async function extractFile(path, destination) {
   if (path.endsWith(".zip")) {
-    const extract = unzipper.Extract({ path: "./bin" });
-
-    fs.createReadStream(path).pipe(extract);
+    return extractFileZip(path, destination)  
   } else if (path.endsWith(".tar.gz")) {
-    const extract = tar.extract("./bin");
-
+    const extract = tar.extract(destination);
     fs.createReadStream(path).pipe(zlib.createGunzip()).pipe(extract);
   }
 
-  return true
+  return true;
 }
 
-async function installBinaries() {
+async function installBinaries(destination) {
   if (!PLATFORM) throw UNSUPPORTED_PLATFORM;
   if (!ARCH) throw UNSUPPORTED_ARCH;
 
@@ -67,12 +69,21 @@ async function installBinaries() {
   const asset = release.assets.filter((a) => regex.test(a.name))[0];
   if (!asset) throw new Error(`Cannot find an asset for ${PLATFORM} - ${ARCH}`);
 
+  const checksumAsset = release.assets.find(a => a.name === "checksums.txt")
   try {
-    console.log("Downloading binaries from github release...")
-    const res = await fetch(asset.browser_download_url);
-    fs.mkdirSync(path.resolve(__dirname, "./bin"), { recursive: true });
+    console.log("Downloading checksums...")
+    const cres = await fetch(checksumAsset.browser_download_url);
+    const checksumText = await cres.text()
+    const checksumList = checksumText.split(/\n/g).map(c => c.split(/\s+/g))
+    const [checksum, checksumFilename] = checksumList.find(c => asset.browser_download_url.includes(c[1]))
 
-    const filepath = path.resolve(__dirname, "./bin", asset.name);
+    console.log("Downloading binaries from github release...");
+    console.log(asset.browser_download_url);
+
+    const res = await fetch(asset.browser_download_url);
+    fs.mkdirSync(destination, { recursive: true });
+
+    const filepath = path.resolve(destination, asset.name);
     const stream = fs.createWriteStream(filepath);
 
     const reader = res.body.getReader();
@@ -91,34 +102,55 @@ async function installBinaries() {
     stream.once("error", (err) => {
       console.log("Error when trying to download the file", err);
     });
-  
+
     stream.once("close", async () => {
-      console.log("Extracting file...")
-      await extractFile(filepath)
+      console.log("Validating file...")
+      const file = fs.readFileSync(filepath)
+      const hash = createHash("sha256").update(file).digest("hex")
+      
+      if (hash != checksum) {
+        throw new Error("The file checksum does not match with checksums.txt")
+      }
+
+      console.log("Extracting file...");
+      await extractFile(filepath, destination);
       fs.unlinkSync(filepath);
 
 
-      console.log("Square Cloud CLI successfuly installed, please perform \"squarecloud --help\"")
+      console.log("Square Cloud CLI successfuly installed, please perform \"squarecloud --help\"");
     });
   } catch (err) {
     console.log("Error when trying to download and extract the file", err);
   }
 }
 
-(async () => {
-  const argv = process.argv;
-  if (argv[2] === "update") {
-    installBinaries();
-    return;
+function getBinDir() {
+  switch (process.platform) {
+    case "linux":
+      return path.join(process.env.HOME, ".squarecloud");
+    case "win32":
+      return path.join(process.env.APPDATA, "squarecloud");
+    default:
+      throw new Error("Platform Unsupported");
   }
+}
 
-  const binDir = path.resolve(__dirname, "./bin");
-  const execfile = path.resolve(binDir, "squarecloud");
+function getExecFile() {
+  const extension = process.platform === "win32" ? ".exe" : "";
+  const binDir = getBinDir();
+  const execfile = path.resolve(binDir, `squarecloud${extension}`);
 
-  if (!fs.existsSync(execfile)) {
-    await installBinaries();
-    console.log("Square Cloud CLI successfuly installed, please perform the command again.")
-    return
+  return execfile;
+}
+
+(async () => {
+  const binDir = getBinDir();
+  const execfile = getExecFile();
+
+  const argv = process.argv;
+  if (argv[2] === "update" || !fs.existsSync(execfile)) {
+    installBinaries(binDir);
+    return;
   }
 
   stdout.write(execFileSync(execfile, argv.slice(2)));
